feat: Add MongoDB connection and collection initialization
This commit adds code to connect to a MongoDB database and initialize a collection for the BlogService. It uses the `go.mongodb.org/mongo-driver` package to establish a connection to the local MongoDB server running on port 27017. The `mydb` database is selected, and the `blog` collection is created. This allows the server to store and retrieve blog data from the MongoDB database.