feat: Implement CreateBlog RPC method in BlogService
This commit adds the implementation for the CreateBlog RPC method in the BlogService. It allows clients to create a new blog by inserting the blog data into the MongoDB collection. The method takes a CreateBlogRequest as input and returns a CreateBlogResponse with the newly created blog. This functionality enables users to add new blogs to the system.